+++ demos/GeometryMeshRenderer.js
@@ -20,6 +20,7 @@
    * @param {GeometryMesh} geometry
    * @param {string} options.textColor
    * @param {ColorSpace2d} options.colorSpace – (optional) With a function getColorAt(x,y,z) => Color
+   * @param {boolean} options.useDepthBuffer – (optional) Endable/disable depth buffer.
    */
   GeometryMeshRenderer.prototype.drawGeometry = function (draw, fill, geometry, options) {
     // var minMax = getMinMax(geometry.vertices);
@@ -82,16 +83,16 @@
           return colorC;
         }
       };
-      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix0, fakeColorSpaceA); // Color.makeRGB(128, 255, 0));
-      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix1, fakeColorSpaceB); // Color.makeRGB(0, 0, 255));
-      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix2, fakeColorSpaceC); // Color.makeRGB(255, 0, 0));
+      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix0, fakeColorSpaceA, options); // Color.makeRGB(128, 255, 0));
+      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix1, fakeColorSpaceB, options); // Color.makeRGB(0, 0, 255));
+      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix2, fakeColorSpaceC, options); // Color.makeRGB(255, 0, 0));
       // Use this on white
       // this.drawGeometry(draw, fill, geometry, minMax, transformMatrix0, Color.makeRGB(128, 0, 255));
       // this.drawGeometry(draw, fill, geometry, minMax, transformMatrix1, Color.makeRGB(255, 255, 0));
       // this.drawGeometry(draw, fill, geometry, minMax, transformMatrix2, Color.makeRGB(0, 255, 255));
     } else if (options.colorSpace) {
       // console.log("Use options.colorSpace");
-      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix0, options.colorSpace);
+      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix0, options.colorSpace, options);
     } else {
       // this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix0, Color.makeRGB(92, 92, 92));
       const edgeColor = Color.makeRGB(192, 192, 192);
@@ -100,7 +101,7 @@
           return edgeColor;
         }
       };
-      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix0, fakeColorSpace); // Color.makeRGB(192, 192, 192));
+      this.drawGeometryEdges(draw, fill, geometry, minMax, transformMatrix0, fakeColorSpace, options); // Color.makeRGB(192, 192, 192));
     }
 
     this.drawGeometryVertices(draw, fill, geometry, transformMatrix0, {
@@ -110,17 +111,59 @@
   };
 
   // +---------------------------------------------------------------------------------
+  // | Transform all vertices and store in a new array.
+  // +-------------------------------
+  GeometryMeshRenderer.prototype.createTransformedVertices = function (geometry, transformMatrix) {
+    var resultVertices = [];
+    var v3_original = null; // Vert3()
+    for (var v = 0; v < geometry.vertices.length; v++) {
+      // a3_original = geometry.vertices[geometry.edges[e][0]];
+      // b3_original = geometry.vertices[geometry.edges[e][1]];
+      v3_original = geometry.vertices[v];
+      var a3_transformed = transformMatrix.apply3(v3_original);
+      resultVertices.push(a3_transformed);
+    }
+    // return resultVertices;
+    return new GeometryMesh(resultVertices, geometry.edges); // COPY ARRAY???
+  };
+
+  // +---------------------------------------------------------------------------------
+  // | Sort edges by by z component (of first vertex)
+  // +-------------------------------
+  GeometryMeshRenderer.prototype.createZBufferVertices = function (geometry) {
+    geometry.edges.sort(function (edgeA, edgeB) {
+      var va = geometry.vertices[edgeA[0]];
+      var vb = geometry.vertices[edgeB[0]];
+      return va.z - vb.z;
+    });
+  };
+
+  // +---------------------------------------------------------------------------------
   // | Draw the edges of a geometry.
   // +-------------------------------
-  GeometryMeshRenderer.prototype.drawGeometryEdges = function (draw, fill, geometry, minMax, transformMatrix, colorSpace) {
+  GeometryMeshRenderer.prototype.drawGeometryEdges = function (
+    draw,
+    fill,
+    geometry,
+    minMax,
+    transformMatrix,
+    colorSpace,
+    options
+  ) {
     var a3_vert = new Vert3();
     var a3_original = null; // Vert3()
     var b3_original = null; // Vert3()
+    var transformedVerticesGeometry = this.createTransformedVertices(geometry, transformMatrix);
+    if (options && options.useDepthBuffer) {
+      this.createZBufferVertices(transformedVerticesGeometry);
+    }
     for (var e in geometry.edges) {
       a3_original = geometry.vertices[geometry.edges[e][0]];
       b3_original = geometry.vertices[geometry.edges[e][1]];
-      var a3 = transformMatrix.apply3(a3_original);
-      var b3 = transformMatrix.apply3(b3_original);
+      // var a3 = transformMatrix.apply3(a3_original);
+      // var b3 = transformMatrix.apply3(b3_original);
+      var a3 = transformedVerticesGeometry.vertices[geometry.edges[e][0]];
+      var b3 = transformedVerticesGeometry.vertices[geometry.edges[e][1]];
 
       var a2 = this.applyProjection(a3);
       var b2 = this.applyProjection(b3);
@@ -133,7 +176,7 @@
       a3_vert.x = a3.x;
       a3_vert.y = a3.y;
       a3_vert.z = a3.z;
-      var midpoint = a3_vert.lerp(b3, 0.5);
+      // var midpoint = a3_vert.lerp(b3, 0.5);
       // var colorObject = colorSpace.getColorAt(midpoint.x, midpoint.y, midpoint.z);
       var colorObject = colorSpace.getColorAt(a3_original.x, a3_original.y, a3_original.z);