+++ src/cjs/utils/WebColors.js
@@ -9,17 +9,17 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.shuffleWebColors = exports.WebColors = exports.LightGreen = exports.Green = exports.Teal = exports.Cyan = exports.LightBlue = exports.Blue = exports.Indigo = exports.DeepPurple = exports.Purple = exports.Pink = exports.Red = void 0;
 var Color_1 = require("./datastructures/Color");
-exports.Red = Color_1.Color.makeRGB(255, 67, 55);
-exports.Pink = Color_1.Color.makeRGB(232, 31, 100);
-exports.Purple = Color_1.Color.makeRGB(156, 39, 175);
-exports.DeepPurple = Color_1.Color.makeRGB(103, 59, 184);
-exports.Indigo = Color_1.Color.makeRGB(64, 81, 181);
-exports.Blue = Color_1.Color.makeRGB(35, 151, 245);
-exports.LightBlue = Color_1.Color.makeRGB(6, 170, 245);
-exports.Cyan = Color_1.Color.makeRGB(3, 189, 214);
-exports.Teal = Color_1.Color.makeRGB(1, 150, 137);
-exports.Green = Color_1.Color.makeRGB(77, 175, 82);
-exports.LightGreen = Color_1.Color.makeRGB(141, 195, 67);
+exports.Red = Color_1.Color.Red; // Color.makeRGB(255,67,55);
+exports.Pink = Color_1.Color.Pink; // Color.makeRGB(232,31,100);
+exports.Purple = Color_1.Color.Purple; // Color.makeRGB(156,39,175);
+exports.DeepPurple = Color_1.Color.DeepPink; // Color.makeRGB(103,59,184);
+exports.Indigo = Color_1.Color.Indigo; // Color.makeRGB(64,81,181);
+exports.Blue = Color_1.Color.Blue; // Color.makeRGB(35,151,245);
+exports.LightBlue = Color_1.Color.LightBlue; // Color.makeRGB(6,170,245);
+exports.Cyan = Color_1.Color.Cyan; // Color.makeRGB(3,189,214);
+exports.Teal = Color_1.Color.Teal; // Color.makeRGB(1,150,137);
+exports.Green = Color_1.Color.Green; // Color.makeRGB(77,175,82);
+exports.LightGreen = Color_1.Color.LightGreen; // Color.makeRGB(141,195,67);
 /**
  * A set of beautiful web colors (I know, beauty is in the eye of the beholder).
  *
@@ -29,19 +29,7 @@ exports.LightGreen = Color_1.Color.makeRGB(141, 195, 67);
  * @requires Color
  *
  */
-exports.WebColors = [
-    exports.Red,
-    exports.Pink,
-    exports.Purple,
-    exports.DeepPurple,
-    exports.Indigo,
-    exports.Blue,
-    exports.LightBlue,
-    exports.Cyan,
-    exports.Teal,
-    exports.Green,
-    exports.LightGreen
-];
+exports.WebColors = [exports.Red, exports.Pink, exports.Purple, exports.DeepPurple, exports.Indigo, exports.Blue, exports.LightBlue, exports.Cyan, exports.Teal, exports.Green, exports.LightGreen];
 /**
  * A helper function to shuffle the colors into a new order.
  */